lab2, bug report and api update...
It's weird that on real rpi3, it may crash if you access mem too fast and the mem addr is far from where the current execution image loaded.
This commit will crash on rpi3 when -O3 is in the compile flag.
And if -O3 presents, adding a delay (such as uart_print("%c",'\0')) can also fix the problem.

Add diy_string.h to replace string.h
Modify .c to fit diy_string.h