lab5: Fix: Possibly corrupt user thread's stack. Add mailbox test
When calling system call, the user thread's stack might be corrupted.
When exception triggered in nested function calls in user thread, it might fail to return to the most outside call.
Don't know if it's a problem with function call or not, but it work well when it's KERNEL thread.
Fixed by allocating another thread space for user thread.

Add mailbox test. (Above bug discovered when adding mailbox test)